requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==3.0.1
docopt==0.6.2
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.32
outcome==1.2.0
packaging==23.0
pipreqs==0.4.11
protobuf==3.20.3
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
yarg==0.1.9
